Move AntiGhostPing cog code from lib/bot/__init__.py to lib/cogs/anti-ghostping.py
