+++ lib/bot/__init__.py
@@ -64,7 +64,6 @@ class BotRoot(commands.Bot):
         super().__init__(
             command_prefix=prefix, intents=intents)
         self.connection = db.DBConnection()
-        self.add_cog(AntiGhostPing(self))
         self.add_cog(Configuration(self))
 
     async def on_ready(self):
@@ -420,98 +419,3 @@ class Configuration(commands.Cog):
         for field in fields:
             embed.add_field(name=field, value=fields[field])
         await direct_message.send(embed=embed)
-
-
-class AntiGhostPing(commands.Cog):
-    """ Listen for and handle ghost pings
-"""
-    def __init__(self, bot):
-        self.bot = bot
-
-    @commands.Cog.listener()
-    async def on_message_delete(self, message):
-        """ Notify logging of event reference
-            Check for and handle host ping
-"""
-        if message.author.bot:
-            return
-        flags = await self.parse(message)
-        if flags:
-            await self.detected(message, flags)
-
-    async def parse(self, message):
-        """ Parse message for all specified mentions
-"""
-        # Get guild preferences from db.sqlite
-        select_preferences_table = """
-        SELECT 
-            roles,
-            members,
-            everyone
-        FROM 
-            preferences
-        WHERE GuildID=?"""
-        columns, prefs = self.bot.connection.execute_query(
-            select_preferences_table, "rr",
-            message.guild.id
-        )[0]
-        preferences = dict(zip(columns, prefs))
-        # Parse message for raw mentions and flags for specified preferences
-        flags = {}
-        # Check for role mentions
-        if preferences["roles"] and message.raw_role_mentions:
-            role_mentions = [
-                discord.utils.get(message.guild.roles, id=i).name
-                for i in message.raw_role_mentions
-            ]
-            flags.setdefault("Roles Mentioned", ", ".join(role_mentions))
-        # Check for member mentions
-        if preferences["members"] and message.raw_mentions:
-            raw_mentions = [
-                discord.utils.get(message.guild.members, id=i).name
-                for i in message.raw_mentions
-            ]
-            flags.setdefault("Members Mentioned", ", ".join(raw_mentions))
-        # Check for everyone mentions
-        if preferences["everyone"] and message.mention_everyone:
-            flags.setdefault("Other Groups Mentioned", "everyone")
-        return flags
-
-    async def detected(self, message, flags):
-        """ Alert guild by sending message to specified channel
-"""
-        # Get guild preferences from db.sqlite
-        select_preferences_table = """
-        SELECT channel
-        FROM preferences
-        WHERE GuildID=?"""
-        columns, prefs = self.bot.connection.execute_query(
-            select_preferences_table, "rr",
-            message.guild.id
-        )[0]
-        preferences = dict(zip(columns, prefs))
-        # Get notification channel preferences from guild preferences
-        if preferences is None:
-            channel = message.channel
-        else:
-            channel = discord.utils.get(
-                message.guild.channels,
-                id=preferences["channel"]
-            )
-            if channel is None:
-                channel = message.channel
-        # Send notifying embed to specified channel
-        embed = discord.Embed(
-            title="Ghost Ping Detected :no_entry_sign: :ghost:",
-            color=0x0000ff
-        )
-        fields = {
-            "Member": message.author.name, "Message": message.content,
-            "Channel": message.channel.name
-        }
-        fields = {**fields, **flags}
-        for field in fields:
-            embed.add_field(name=field, value=fields[field])
-        embed.set_footer(
-            text=f"Detect At: {message.created_at.strftime('%D %T')}")
-        await channel.send(embed=embed)
